改用mongo储存实时数据

+++ common/stockOptions.go
@@ -11,7 +11,7 @@ type RankOpt struct {
 	MarketType string // 市场类型
 	SortName   string
 	Sorted     bool //排序
-	Page       int
+	Page       int64
 }
 
 // SearchOpt 搜索
